Move templateUtils module
