+++ ui-participant/src/landing/sections/HeroCenteredTemplate.tsx
@@ -5,12 +5,13 @@ import ReactMarkdown from 'react-markdown'
 
 import { SectionConfig } from 'api/api'
 import { getSectionStyle } from 'util/styleUtils'
-import { TemplateComponentProps } from 'util/templateUtils'
 import { withValidatedSectionConfig } from 'util/withValidatedSectionConfig'
 import { requireOptionalArray, requireOptionalString } from 'util/validationUtils'
 
 import ConfiguredButton, { ButtonConfig, validateButtonConfig } from '../ConfiguredButton'
 import PearlImage, { PearlImageConfig, validatePearlImageConfig } from '../PearlImage'
+
+import { TemplateComponentProps } from './templateUtils'
 
 type HeroCenteredTemplateConfig = {
   blurb?: string, //  text below the title
