+++ ui-participant/src/landing/sections/templateUtils.ts
